Any profile page now always viewable
Add feature for any profile page to be viewable by any user, including a user who is not signed in. If the user viewing a profile page is not the owner of that profile page, then the ability for editing the user profile is disabled. User can only edit their own profile.